@@ -3,29 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^13.1.8",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^15.0.1",
-    "@types/qs": "^6.9.6",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "@types/react-router-dom": "^5.1.7",
-    "autoprefixer": "^10.2.5",
-    "date-fns": "^2.21.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.1.1",
+    "@types/jest": "^27.5.0",
+    "@types/node": "^16.11.33",
+    "@types/qs": "^6.9.7",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "autoprefixer": "^10.4.7",
+    "date-fns": "^2.28.0",
     "fetch-mock": "^9.11.0",
-    "node-fetch": "^2.6.1",
-    "postcss-loader": "^4.2.0",
-    "qs": "^6.10.1",
-    "react": "^17.0.2",
-    "react-app-rewired": "^2.1.8",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^6.15.7",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "tailwindcss": "^2.1.2",
-    "typescript": "^4.2.4"
+    "node-fetch": "^2.6.7",
+    "postcss-loader": "^6.2.1",
+    "prettier": "^2.6.2",
+    "qs": "^6.10.3",
+    "react": "^18.1.0",
+    "react-app-rewired": "^2.2.1",
+    "react-dom": "^18.1.0",
+    "react-hook-form": "^7.30.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "tailwindcss": "^3.0.24",
+    "typescript": "^4.6.4"
   },
   "browserslist": {
     "production": [
@@ -51,9 +51,17 @@
       "/src/index.tsx"
     ]
   },
+  "prettier": {
+    "printWidth": 120,
+    "tabWidth": 4,
+    "singleQuote": true,
+    "trailingComma": "all"
+  },
   "scripts": {
-    "start": "react-app-rewired start",
     "build": "react-app-rewired build",
+    "cs-fix": "prettier --write src",
+    "cs": "prettier --check src",
+    "start": "react-app-rewired start",
     "test": "react-app-rewired test"
   }
 }
