@@ -1,6 +1,6 @@
 import { FC, useState } from 'react';
 import { CreatePet } from '../../../ApiClient/Pet';
-import { Link, useHistory } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import HttpError from '../../../Model/Error/HttpError';
 import HttpErrorPartial from '../../Partial/HttpError';
 import PetForm from '../../Form/PetForm';
@@ -8,8 +8,7 @@ import UnprocessableEntity from '../../../Model/Error/UnprocessableEntity';
 import PetRequest from '../../../Model/Pet/PetRequest';
 
 const Create: FC = () => {
-
-    const history = useHistory();
+    const navigate = useNavigate();
 
     const [httpError, setHttpError] = useState<HttpError>();
 
@@ -23,18 +22,21 @@ const Create: FC = () => {
         } else {
             setHttpError(undefined);
 
-            history.push('/pet');
+            navigate('/pet');
         }
     };
 
     return (
-        <div data-testid='page-pet-create'>
-            {httpError ? (
-                <HttpErrorPartial httpError={httpError} />
-            ) : ''}
+        <div data-testid="page-pet-create">
+            {httpError ? <HttpErrorPartial httpError={httpError} /> : null}
             <h1>Create Pet</h1>
-            <PetForm submitPet={submitPet} unprocessableEntity={httpError instanceof UnprocessableEntity ? httpError : undefined} />
-            <Link to='/pet' className='btn-gray'>List</Link>
+            <PetForm
+                submitPet={submitPet}
+                unprocessableEntity={httpError instanceof UnprocessableEntity ? httpError : undefined}
+            />
+            <Link to="/pet" className="btn-gray">
+                List
+            </Link>
         </div>
     );
 };
