@@ -1,5 +1,5 @@
 import { FC, useState, useEffect } from 'react';
-import { Link, RouteComponentProps, useHistory } from 'react-router-dom';
+import { Link, useNavigate, useParams } from 'react-router-dom';
 import { ReadPet, UpdatePet } from '../../../ApiClient/Pet';
 import HttpError from '../../../Model/Error/HttpError';
 import HttpErrorPartial from '../../Partial/HttpError';
@@ -8,13 +8,11 @@ import PetForm from '../../Form/PetForm';
 import UnprocessableEntity from '../../../Model/Error/UnprocessableEntity';
 import PetRequest from '../../../Model/Pet/PetRequest';
 
-type Props = RouteComponentProps<{ id: string; }>;
+const Update: FC = () => {
+    const params = useParams();
+    const id = params.id as string;
 
-const Update: FC<Props> = ({ match }: Props) => {
-
-    const id = match.params.id;
-
-    const history = useHistory();
+    const navigate = useNavigate();
 
     const [pet, setPet] = useState<PetResponse>();
     const [httpError, setHttpError] = useState<HttpError>();
@@ -45,27 +43,30 @@ const Update: FC<Props> = ({ match }: Props) => {
             setHttpError(undefined);
             setPet(response);
 
-            history.push('/pet');
+            navigate('/pet');
         }
     };
 
     if (!pet && !httpError) {
-        return (<div></div>);
+        return <div></div>;
     }
 
     return (
-        <div data-testid='page-pet-update'>
-            {httpError ? (
-                <HttpErrorPartial httpError={httpError} />
-            ) : ''}
+        <div data-testid="page-pet-update">
+            {httpError ? <HttpErrorPartial httpError={httpError} /> : null}
             <h1>Update Pet</h1>
             {pet ? (
-                <PetForm submitPet={submitPet} defaultPet={pet} unprocessableEntity={httpError instanceof UnprocessableEntity ? httpError : undefined} />
-            ) : ''}
-            <Link to='/pet' className='btn-gray'>List</Link>
+                <PetForm
+                    submitPet={submitPet}
+                    defaultPet={pet}
+                    unprocessableEntity={httpError instanceof UnprocessableEntity ? httpError : undefined}
+                />
+            ) : null}
+            <Link to="/pet" className="btn-gray">
+                List
+            </Link>
         </div>
     );
 };
 
 export default Update;
-
