@@ -5,7 +5,11 @@ import PetFilterFormProps from './PetFilterFormProps';
 import PetFilters from '../../Model/Pet/PetFilters';
 import TextField from './TextField';
 
-const PetFilterForm: FC<PetFilterFormProps> = ({ submitPetFilter, defaultPetFilters: filters, badRequest: error }: PetFilterFormProps) => {
+const PetFilterForm: FC<PetFilterFormProps> = ({
+    submitPetFilter,
+    defaultPetFilters: filters,
+    badRequest: error,
+}: PetFilterFormProps) => {
     const invalidParameterByNameDenormalized = InvalidParameterByNameDenormalizer(error ? error.invalidParameters : []);
 
     const { register, handleSubmit } = useForm<PetFilters>({ defaultValues: filters });
@@ -21,8 +25,15 @@ const PetFilterForm: FC<PetFilterFormProps> = ({ submitPetFilter, defaultPetFilt
     return (
         <form onSubmit={handleSubmit(onSubmit)}>
             <fieldset>
-                <TextField register={register} name='name' label='Name' invalidParameters={invalidParameterByNameDenormalized.get('name') ?? []} />
-                <button data-testid='submit-pet-filter' className='btn-blue'>Filter</button>
+                <TextField
+                    register={register}
+                    name="name"
+                    label="Name"
+                    invalidParameters={invalidParameterByNameDenormalized.get('name') ?? []}
+                />
+                <button data-testid="submit-pet-filter" className="btn-blue">
+                    Filter
+                </button>
             </fieldset>
         </form>
     );
