@@ -9,17 +9,19 @@ type Props = {
 
 const HttpError: FC<Props> = ({ httpError }: Props) => {
     return (
-        <div id='httpError'>
+        <div id="httpError">
             <p>{httpError.title}</p>
-            {httpError.detail ? (<p>{httpError.detail}</p>) : ''}
-            {httpError.instance ? (<p>{httpError.instance}</p>) : ''}
+            {httpError.detail ? <p>{httpError.detail}</p> : null}
+            {httpError.instance ? <p>{httpError.instance}</p> : null}
             {httpError instanceof HttpErrorWithInvalidParameters && httpError.invalidParameters.length > 0 ? (
                 <ul>
                     {httpError.invalidParameters.map((invalidParameter: InvalidParameter, i) => (
-                        <li key={i}><strong>{invalidParameter.name}</strong>: {invalidParameter.reason}</li>
+                        <li key={i}>
+                            <strong>{invalidParameter.name}</strong>: {invalidParameter.reason}
+                        </li>
                     ))}
                 </ul>
-            ) : ''}
+            ) : null}
         </div>
     );
 };
