@@ -2,8 +2,8 @@ import { FC } from 'react';
 import PaginationProps from './PaginationProps';
 
 const Pagination: FC<PaginationProps> = ({ submitPage, currentPage, totalPages, maxPages }: PaginationProps) => {
-    if (totalPages <= 1 || maxPages <=1) {
-        return (<div></div>);
+    if (totalPages <= 1 || maxPages <= 1) {
+        return <div></div>;
     }
 
     const pages = [currentPage];
@@ -28,22 +28,63 @@ const Pagination: FC<PaginationProps> = ({ submitPage, currentPage, totalPages, 
     pages.sort((a, b) => a - b);
 
     return (
-        <ul className='pagination'>
+        <ul className="pagination">
             {currentPage > 2 ? (
-                <li><button onClick={() => { submitPage(1); }}>&laquo;</button></li>
-            ) : ''}
+                <li>
+                    <button
+                        onClick={() => {
+                            submitPage(1);
+                        }}
+                    >
+                        &laquo;
+                    </button>
+                </li>
+            ) : null}
             {currentPage > 1 ? (
-                <li><button onClick={() => { submitPage(currentPage - 1); }}>&lt;</button></li>
-            ) : ''}
+                <li>
+                    <button
+                        onClick={() => {
+                            submitPage(currentPage - 1);
+                        }}
+                    >
+                        &lt;
+                    </button>
+                </li>
+            ) : null}
             {pages.map((page: number) => (
-                <li key={page}><button className={currentPage === page ? 'current' : ''} onClick={() => { submitPage(page); }}>{page}</button></li>
+                <li key={page}>
+                    <button
+                        className={currentPage === page ? 'current' : ''}
+                        onClick={() => {
+                            submitPage(page);
+                        }}
+                    >
+                        {page}
+                    </button>
+                </li>
             ))}
             {currentPage < totalPages ? (
-                <li><button onClick={() => { submitPage(currentPage + 1); }}>&gt;</button></li>
-            ) : ''}
+                <li>
+                    <button
+                        onClick={() => {
+                            submitPage(currentPage + 1);
+                        }}
+                    >
+                        &gt;
+                    </button>
+                </li>
+            ) : null}
             {currentPage < totalPages - 1 ? (
-                <li><button onClick={() => { submitPage(totalPages); }}>&raquo;</button></li>
-            ) : ''}
+                <li>
+                    <button
+                        onClick={() => {
+                            submitPage(totalPages);
+                        }}
+                    >
+                        &raquo;
+                    </button>
+                </li>
+            ) : null}
         </ul>
     );
 };
