@@ -1,6 +1,8 @@
 import InvalidParameter from '../Model/Error/InvalidParameter';
 
-const InvalidParameterByNameDenormalizer = (invalidParameters: Array<InvalidParameter>): Map<string, Array<InvalidParameter>> => {
+const InvalidParameterByNameDenormalizer = (
+    invalidParameters: Array<InvalidParameter>,
+): Map<string, Array<InvalidParameter>> => {
     const errorsByFields = new Map<string, Array<InvalidParameter>>();
     invalidParameters.forEach((invalidParameter: InvalidParameter) => {
         if (!errorsByFields.has(invalidParameter.name)) {
