@@ -1,6 +1,5 @@
 import HttpErrorWithInvalidParameters from './HttpErrorWithInvalidParameters';
 
-class BadRequest extends HttpErrorWithInvalidParameters {
-}
+class BadRequest extends HttpErrorWithInvalidParameters {}
 
 export default BadRequest;
