@@ -1,6 +1,5 @@
 import HttpError from './HttpError';
 
-class InternalServerError extends HttpError {
-}
+class InternalServerError extends HttpError {}
 
 export default InternalServerError;
