@@ -2,6 +2,6 @@ interface InvalidParameter {
     name: string;
     reason: string;
     details?: Object;
-};
+}
 
 export default InvalidParameter;
