@@ -1,6 +1,5 @@
 import HttpError from './HttpError';
 
-class NetworkError extends HttpError {
-}
+class NetworkError extends HttpError {}
 
 export default NetworkError;
