@@ -1,6 +1,5 @@
 import HttpError from './HttpError';
 
-class NotFound extends HttpError {
-}
+class NotFound extends HttpError {}
 
 export default NotFound;
