@@ -1,6 +1,5 @@
 import HttpErrorWithInvalidParameters from './HttpErrorWithInvalidParameters';
 
-class UnprocessableEntity extends HttpErrorWithInvalidParameters {
-}
+class UnprocessableEntity extends HttpErrorWithInvalidParameters {}
 
 export default UnprocessableEntity;
