@@ -2,6 +2,6 @@ import Link from './Link';
 
 interface Links {
     [id: string]: Link;
-};
+}
 
 export default Links;
