@@ -7,7 +7,23 @@ class PetResponse extends PetRequest {
     createdAt: string;
     updatedAt?: string;
     _links: Links;
-    constructor({ id, createdAt, updatedAt, name, tag, vaccinations, _links }: { id: string, createdAt: string, updatedAt?: string, name: string, tag?: string, vaccinations?: Array<Vaccination>, _links?: Links; }) {
+    constructor({
+        id,
+        createdAt,
+        updatedAt,
+        name,
+        tag,
+        vaccinations,
+        _links,
+    }: {
+        id: string;
+        createdAt: string;
+        updatedAt?: string;
+        name: string;
+        tag?: string;
+        vaccinations?: Array<Vaccination>;
+        _links?: Links;
+    }) {
         super({ name, tag, vaccinations });
         this.id = id;
         this.createdAt = createdAt;
