@@ -1,20 +1,19 @@
-import React from 'react';
 import { render } from '@testing-library/react';
 import TextField from '../../../Component/Form/TextField';
 import InvalidParameter from '../../../Model/Error/InvalidParameter';
+import { UseFormRegisterReturn } from 'react-hook-form';
 
 test('default', () => {
-    const register = () => {};
+    const register = (name: string) => ({ ...({} as UseFormRegisterReturn), name });
 
-    const invalidParameters: Array<InvalidParameter> = [
-        { name: 'name', reason: 'Should not be empty' }
-    ];
+    const invalidParameters: Array<InvalidParameter> = [{ name: 'name', reason: 'Should not be empty' }];
 
     const { container } = render(
-        <TextField register={register} name='name' label='label' invalidParameters={invalidParameters} />
+        <TextField register={register} name="name" label="label" invalidParameters={invalidParameters} />,
     );
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div class="form-field error">
                 <label>label</label>
@@ -24,5 +23,8 @@ test('default', () => {
                 </ul>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
