@@ -1,23 +1,26 @@
-import React from 'react';
 import { createMemoryHistory } from 'history';
 import { render } from '@testing-library/react';
-import { Router } from 'react-router-dom';
+import { unstable_HistoryRouter as HistoryRouter } from 'react-router-dom';
 import Home from '../../../Component/Page/Home';
 
 test('default', () => {
     const history = createMemoryHistory();
 
     const { container } = render(
-        <Router history={history}>
+        <HistoryRouter history={history}>
             <Home />
-        </Router>
+        </HistoryRouter>,
     );
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div>
                 <h1>Home</h1>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
