@@ -1,26 +1,25 @@
-import React from 'react';
 import { createMemoryHistory } from 'history';
 import { render, screen } from '@testing-library/react';
-import { Router } from 'react-router-dom';
+import { unstable_HistoryRouter as HistoryRouter } from 'react-router-dom';
 import HttpError from '../../../../Model/Error/HttpError';
 import NotFound from '../../../../Model/Error/NotFound';
 import PetResponse from '../../../../Model/Pet/PetResponse';
 import Read from '../../../../Component/Page/Pet/Read';
 import Vaccination from '../../../../Model/Pet/Vaccination';
 
-let mockReadPet = (id: string) => { };
+let mockReadPet = (id: string) => {};
 
 jest.mock('../../../../ApiClient/Pet', () => {
     return {
         ReadPet: (id: string) => {
             return mockReadPet(id);
-        }
+        },
     };
 });
 
 jest.mock('../../../../Component/Partial/HttpError', () => {
-    return ({ httpError }: { httpError: HttpError; }) => {
-        return (<div>httpError: {httpError.title}</div>);
+    return ({ httpError }: { httpError: HttpError }) => {
+        return <div>httpError: {httpError.title}</div>;
     };
 });
 
@@ -30,39 +29,35 @@ test('not found', async () => {
     };
 
     const history = createMemoryHistory();
-
-    const match = {
-        params: {
-            id: '4d783b77-eb09-4603-b99b-f590b605eaa9'
-        },
-        isExact: true,
-        path: '/pet/:id',
-        url: ''
-    };
+    history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const { container } = render(
-        <Router history={history}>
-            <Read history={history} location={history.location} match={match} />
-        </Router>
+        <HistoryRouter history={history}>
+            <Read />
+        </HistoryRouter>,
     );
 
     await screen.findByTestId('page-pet-read');
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div data-testid="page-pet-read">
                 <div>httpError: title</div>
                 <h1>Read Pet</h1>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('minimal', async () => {
     const pet = new PetResponse({
         id: '4d783b77-eb09-4603-b99b-f590b605eaa9',
         createdAt: '2005-08-15T15:52:01+00:00',
-        name: 'Brownie'
+        name: 'Brownie',
     });
 
     mockReadPet = async (id: string) => {
@@ -70,25 +65,18 @@ test('minimal', async () => {
     };
 
     const history = createMemoryHistory();
-
-    const match = {
-        params: {
-            id: '4d783b77-eb09-4603-b99b-f590b605eaa9'
-        },
-        isExact: true,
-        path: '/pet/:id',
-        url: ''
-    };
+    history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const { container } = render(
-        <Router history={history}>
-            <Read history={history} location={history.location} match={match} />
-        </Router>
+        <HistoryRouter history={history}>
+            <Read />
+        </HistoryRouter>,
     );
 
     await screen.findByTestId('page-pet-read');
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div data-testid="page-pet-read">
                 <h1>Read Pet</h1>
@@ -111,7 +99,10 @@ test('minimal', async () => {
                 </div>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('maximal', async () => {
@@ -121,9 +112,7 @@ test('maximal', async () => {
         updatedAt: '2005-08-15T15:55:01+00:00',
         name: 'Brownie',
         tag: '0001-000',
-        vaccinations: [
-            new Vaccination({ name: 'Rabies' })
-        ]
+        vaccinations: [new Vaccination({ name: 'Rabies' })],
     });
 
     mockReadPet = async (id: string) => {
@@ -131,24 +120,18 @@ test('maximal', async () => {
     };
 
     const history = createMemoryHistory();
-
-    const match = {
-        params: {
-            id: '4d783b77-eb09-4603-b99b-f590b605eaa9'
-        },
-        isExact: true,
-        path: '/pet/:id',
-        url: ''
-    };
+    history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const { container } = render(
-        <Router history={history}>
-            <Read history={history} location={history.location} match={match} />
-        </Router>
+        <HistoryRouter history={history}>
+            <Read />
+        </HistoryRouter>,
     );
+
     await screen.findByTestId('page-pet-read');
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div data-testid="page-pet-read">
                 <h1>Read Pet</h1>
@@ -175,5 +158,8 @@ test('maximal', async () => {
                 </div>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
