@@ -1,7 +1,6 @@
-import React from 'react';
 import { createMemoryHistory } from 'history';
 import { render, screen } from '@testing-library/react';
-import { Router } from 'react-router-dom';
+import { unstable_HistoryRouter as HistoryRouter } from 'react-router-dom';
 import HttpError from '../../../../Model/Error/HttpError';
 import NotFound from '../../../../Model/Error/NotFound';
 import PetFormProps from '../../../../Component/Form/PetFormProps';
@@ -12,8 +11,8 @@ import Update from '../../../../Component/Page/Pet/Update';
 import Vaccination from '../../../../Model/Pet/Vaccination';
 import userEvent from '@testing-library/user-event';
 
-let mockReadPet = (id: string) => { };
-let mockUpdatePet = (id: string, pet: PetRequest) => { };
+let mockReadPet = (id: string) => {};
+let mockUpdatePet = (id: string, pet: PetRequest) => {};
 
 jest.mock('../../../../ApiClient/Pet', () => {
     return {
@@ -22,23 +21,23 @@ jest.mock('../../../../ApiClient/Pet', () => {
         },
         UpdatePet: (id: string, pet: PetRequest) => {
             return mockUpdatePet(id, pet);
-        }
+        },
     };
 });
 
 jest.mock('../../../../Component/Form/PetForm', () => {
     return ({ submitPet }: PetFormProps) => {
         const onSubmit = () => {
-            submitPet({ name: 'Brownie' });
+            submitPet({ ...({} as PetRequest), name: 'Brownie' });
         };
 
-        return (<button data-testid="test-button" onClick={onSubmit}></button>);
+        return <button data-testid="test-button" onClick={onSubmit}></button>;
     };
 });
 
 jest.mock('../../../../Component/Partial/HttpError', () => {
-    return ({ httpError }: { httpError: HttpError; }) => {
-        return (<div>httpError: {httpError.title}</div>);
+    return ({ httpError }: { httpError: HttpError }) => {
+        return <div>httpError: {httpError.title}</div>;
     };
 });
 
@@ -48,25 +47,18 @@ test('not found', async () => {
     };
 
     const history = createMemoryHistory();
-
-    const match = {
-        params: {
-            id: '4d783b77-eb09-4603-b99b-f590b605eaa9'
-        },
-        isExact: true,
-        path: '/pet/:id',
-        url: ''
-    };
+    history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const { container } = render(
-        <Router history={history}>
-            <Update history={history} location={history.location} match={match} />
-        </Router>
+        <HistoryRouter history={history}>
+            <Update />
+        </HistoryRouter>,
     );
 
     await screen.findByTestId('page-pet-update');
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div data-testid="page-pet-update">
                 <div>httpError: title</div>
@@ -74,14 +66,17 @@ test('not found', async () => {
                 <a class="btn-gray" href="/pet">List</a>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('minimal', async () => {
     const pet = new PetResponse({
         id: '4d783b77-eb09-4603-b99b-f590b605eaa9',
         createdAt: '2005-08-15T15:52:01+00:00',
-        name: 'Brownie'
+        name: 'Brownie',
     });
 
     mockReadPet = async (id: string) => {
@@ -89,25 +84,18 @@ test('minimal', async () => {
     };
 
     const history = createMemoryHistory();
-
-    const match = {
-        params: {
-            id: '4d783b77-eb09-4603-b99b-f590b605eaa9'
-        },
-        isExact: true,
-        path: '/pet/:id',
-        url: ''
-    };
+    history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const { container } = render(
-        <Router history={history}>
-            <Update history={history} location={history.location} match={match} />
-        </Router>
+        <HistoryRouter history={history}>
+            <Update />
+        </HistoryRouter>,
     );
 
     await screen.findByTestId('page-pet-update');
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div data-testid="page-pet-update">
                 <h1>Update Pet</h1>
@@ -115,7 +103,10 @@ test('minimal', async () => {
                 <a class="btn-gray" href="/pet">List</a>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('unprocessable entity', async () => {
@@ -124,9 +115,7 @@ test('unprocessable entity', async () => {
         createdAt: '2005-08-15T15:52:01+00:00',
         updatedAt: '2005-08-15T15:55:01+00:00',
         name: 'Brownie',
-        vaccinations: [
-            new Vaccination({ name: 'Rabies' })
-        ]
+        vaccinations: [new Vaccination({ name: 'Rabies' })],
     });
 
     mockReadPet = async (id: string) => {
@@ -138,29 +127,24 @@ test('unprocessable entity', async () => {
     };
 
     const history = createMemoryHistory();
-
-    const match = {
-        params: {
-            id: '4d783b77-eb09-4603-b99b-f590b605eaa9'
-        },
-        isExact: true,
-        path: '/pet/:id',
-        url: ''
-    };
+    history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const { container } = render(
-        <Router history={history}>
-            <Update history={history} location={history.location} match={match} />
-        </Router>
+        <HistoryRouter history={history}>
+            <Update />
+        </HistoryRouter>,
     );
+
+    await screen.findByTestId('page-pet-update');
 
     const testButton = await screen.findByTestId('test-button');
 
-    userEvent.click(testButton);
+    await userEvent.click(testButton);
 
-    await screen.findByTestId('test-button');
+    await screen.findByText(/httpError/);
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div data-testid="page-pet-update">
                 <div>httpError: title</div>
@@ -169,7 +153,10 @@ test('unprocessable entity', async () => {
                 <a class="btn-gray" href="/pet">List</a>
             </div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('successful', async () => {
@@ -178,9 +165,7 @@ test('successful', async () => {
         createdAt: '2005-08-15T15:52:01+00:00',
         updatedAt: '2005-08-15T15:55:01+00:00',
         name: 'Brownie',
-        vaccinations: [
-            new Vaccination({ name: 'Rabies' })
-        ]
+        vaccinations: [new Vaccination({ name: 'Rabies' })],
     });
 
     mockReadPet = async (id: string) => {
@@ -192,29 +177,21 @@ test('successful', async () => {
     };
 
     const history = createMemoryHistory();
-
-    const match = {
-        params: {
-            id: '4d783b77-eb09-4603-b99b-f590b605eaa9'
-        },
-        isExact: true,
-        path: '/pet/:id',
-        url: ''
-    };
+    history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const { container } = render(
-        <Router history={history}>
-            <Update history={history} location={history.location} match={match} />
-        </Router>
+        <HistoryRouter history={history}>
+            <Update />
+        </HistoryRouter>,
     );
 
-    expect(history.location.pathname).toBe('/');
+    await screen.findByTestId('page-pet-update');
+
+    expect(history.location.pathname).toBe('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
 
     const testButton = await screen.findByTestId('test-button');
 
-    userEvent.click(testButton);
-
-    await screen.findByTestId('test-button');
+    await userEvent.click(testButton);
 
     expect(history.location.pathname).toBe('/pet');
 });
