@@ -1,44 +1,46 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
 import Pagination from '../../../Component/Partial/Pagination';
 import userEvent from '@testing-library/user-event';
 
 test('max pages 1', () => {
-    const submitPage = (page: number): void => {
-
-    };
+    const submitPage = (page: number): void => {};
 
     const { container } = render(<Pagination currentPage={1} maxPages={1} totalPages={10} submitPage={submitPage} />);
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div></div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('total pages 1', () => {
-    const submitPage = (page: number): void => {
-
-    };
+    const submitPage = (page: number): void => {};
 
     const { container } = render(<Pagination currentPage={1} maxPages={7} totalPages={1} submitPage={submitPage} />);
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <div></div>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('current 1', () => {
-    const submitPage = (page: number): void => {
-
-    };
+    const submitPage = (page: number): void => {};
 
     const { container } = render(<Pagination currentPage={1} maxPages={7} totalPages={10} submitPage={submitPage} />);
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <ul class="pagination">
                 <li><button class="current">1</button></li>
@@ -52,17 +54,19 @@ test('current 1', () => {
                 <li><button>»</button></li>
             </ul>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('current 4', () => {
-    const submitPage = (page: number): void => {
-
-    };
+    const submitPage = (page: number): void => {};
 
     const { container } = render(<Pagination currentPage={4} maxPages={7} totalPages={10} submitPage={submitPage} />);
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <ul class="pagination">
                 <li><button>«</button></li>
@@ -78,17 +82,19 @@ test('current 4', () => {
                 <li><button>»</button></li>
             </ul>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('current 7', () => {
-    const submitPage = (page: number): void => {
-
-    };
+    const submitPage = (page: number): void => {};
 
     const { container } = render(<Pagination currentPage={7} maxPages={7} totalPages={10} submitPage={submitPage} />);
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <ul class="pagination">
                 <li><button>«</button></li>
@@ -104,17 +110,19 @@ test('current 7', () => {
                 <li><button>»</button></li>
             </ul>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('current 10', () => {
-    const submitPage = (page: number): void => {
-
-    };
+    const submitPage = (page: number): void => {};
 
     const { container } = render(<Pagination currentPage={10} maxPages={7} totalPages={10} submitPage={submitPage} />);
 
-    expect(container.outerHTML).toBe(`
+    expect(container.outerHTML).toBe(
+        `
         <div>
             <ul class="pagination">
                 <li><button>«</button></li>
@@ -128,7 +136,10 @@ test('current 10', () => {
                 <li><button class="current">10</button></li>
             </ul>
         </div>
-    `.replace(/\n/g, '').replace(/ {2,}/g, ''));
+    `
+            .replace(/\n/g, '')
+            .replace(/ {2,}/g, ''),
+    );
 });
 
 test('buttons', async () => {
@@ -140,8 +151,8 @@ test('buttons', async () => {
 
     render(<Pagination currentPage={7} maxPages={7} totalPages={10} submitPage={submitPage} />);
 
-    for (const element of screen.getAllByRole('button')) {
-        userEvent.click(element);
+    for await (const element of screen.getAllByRole('button')) {
+        await userEvent.click(element);
     }
 
     expect(pages).toEqual([1, 6, 4, 5, 6, 7, 8, 9, 10, 8, 10]);
