@@ -5,7 +5,7 @@ test('default', () => {
     const invalidParameters: Array<InvalidParameter> = [
         { name: 'name', reason: 'wrong type', details: { key: 'value1' } },
         { name: 'name', reason: 'not empty', details: { key: 'value2' } },
-        { name: 'description', reason: 'to long', details: { key: 'value3' } }
+        { name: 'description', reason: 'to long', details: { key: 'value3' } },
     ];
 
     const invalidParameterByName = InvalidParameterByNameDenormalizer(invalidParameters);
@@ -25,5 +25,7 @@ test('default', () => {
     expect(InvalidParametersOfDescription).toBeInstanceOf(Array);
     expect(InvalidParametersOfDescription).toHaveLength(1);
 
-    expect(InvalidParametersOfDescription instanceof Array ? InvalidParametersOfDescription[0] : null).toBe(invalidParameters[2]);
+    expect(InvalidParametersOfDescription instanceof Array ? InvalidParametersOfDescription[0] : null).toBe(
+        invalidParameters[2],
+    );
 });
